Move Seeder to `rakelib` folder
